\name{bimd-package}
\alias{bimd-package}
\alias{bimd}

\docType{package}

\title{Belgian Index of Multiple Deprivation}

\description{
The \pkg{bimd} package provides datasets containing the scores, ranks and deciles of the Belgian Index of Multiple Deprivation and its subdomains, by statistical sector and municipality.
}

\details{
  \tabular{ll}{
    Package:    \tab  bimd\cr
    Type:       \tab  Package\cr
    Version:    \tab  0.0.1\cr
    Date:       \tab  2024-07-18\cr
    Authors:    \tab  Brecht Devleesschauwer\cr
    Maintainer: \tab  Brecht Devleesschauwer <\email{brechtdv@gmail.com}>\cr
    Depends:    \tab  R (>= 4.3.0)\cr
    License:    \tab  GPL 3\cr
}

Available datasets in the \pkg{bimd} package:
  \tabular{ll}{
    \code{bimd2011_sect}	\tab BIMD2011 by statistical sector.\cr
    \code{bimd2011_muni}	\tab BIMD2011 by municipality.\cr\cr
    \code{bimd2001_sect}	\tab BIMD2001 by statistical sector.\cr
    \code{bimd2001_muni}	\tab BIMD2001 by municipality.\cr\cr
  }
}

\keyword{package}
